add folder controller
