+++ be/src/routes/employeeManager.js
@@ -1,56 +1,8 @@
 // quản lí nhân sự
 const express = require("express");
-const PersonalInformation = require("../entities/PersonalInformation");
-const ContractInformation = require("../entities/ContractInformation");
+const employeeManagerHandler = require("../controller/employeeManagerController")
 const router = express.Router();
 
-router.get('/', async (req, res) => {
-    try {
-        const listPerson = await PersonalInformation.find();
-        if (!listPerson) {
-            return res.status(404).json({
-                error: true,
-                msg: 'Not Found',
-                success: false,
-            })
-        }
-        let arr = [];
-        for (const person of listPerson) {
-            console.log('person', person)
-            arr.push(await ContractInformation.findOne({ username: person.username }));
-        }
-        if (arr.length == 0) {
-            return res.status(404).json({
-                error: true,
-                msg: 'No contractInfo user',
-                success: false,
-            })
-        }
-        const contractInfo = [];
-        for(let i = 0; i < arr.length; i++) {
-            console.log('...listPerson[i]',listPerson[i]);
-            let a = {
-                id: i,
-                fullname: listPerson[i].fullname,
-                username: listPerson[i].username,
-                department: listPerson[i].department,
-                nameContract: arr[i].nameContract,
-            }
-            contractInfo.push(a);
-        }
-        return res.status(200).json({
-            error: false,
-            listPerson: contractInfo,
-            msg: 'Successfully',
-            success: true,
-        })
-    } catch (error) {
-        return res.status(404).json({
-            error: true,
-            msg: 'Error: ' + error.message,
-            success: false,
-        })
-    }
-})
+router.get('/', employeeManagerHandler.handleGetEmployeeManager);
 
 module.exports = router;
